requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
llama-cpp-python
numpy
pandas
scikit-learn
python-dotenv
pydantic
requests